fix(paged_attn): implement correct right-bottom causal mask for chunked prefill
This commit fixes a critical bug in the PagedAttention CUDA kernel where the
causal mask was not properly aligned for chunked prefill scenarios (q_len != kv_len).

Changes:
- cu/cpp: Pass `positions` tensor to the CUDA kernel to fetch the absolute
  position of each query token using its flattened index (`gy`).
- cu/cpp: Update the inner loop condition to `gx <= position` to enforce
  strict right-bottom causality. Added necessary dtype and shape checks.
- python(test): Replace the buggy `is_causal=True` in PyTorch SDPA with a
  custom `gen_right_bottom_mask` to generate the correct reference attention.
- python(engine): Cast `positions` tensor to `torch.int32` in `ModelRunner`
  to match the C++ backend expectation.
- docs: Add explicit warnings in README about `gy` vs `seq_idx` and the
  hidden dangers of using `is_causal=True` in `F.scaled_dot_product_attention`.

+++ femtovllm/engine/model_runner.py
@@ -71,7 +71,7 @@ class ModelRunner:
         )
         positions = torch.tensor(
             positions,
-            dtype=torch.long,
+            dtype=torch.int32,
             device=self.device,
         )
         cu_seqlens = torch.tensor(
